+++ npm/designsystem/src/components/Duolist/index.ts
@@ -1,0 +1,3 @@
+import Duolist from './Duolist';
+export * from './Duolist';
+export default Duolist;